+++ src/app/validators/checkUserId.ts
@@ -1,6 +1,6 @@
 import { validate as validateUuid } from 'uuid';
 import * as database from '../database.json';
-import { StatusCodes, ErrorMessages } from '../types';
+import { StatusCodes, ErrorMessages } from '../utils/types';
 
 const checkUserId = (userId: string) => {
     if (!validateUuid(userId)) {
